Meshes can now also be saved, and now support normals and uvs. Added obj2mesh CLI tool for converting .obj file to .mesh files.

+++ src/SlimEngine/viewport/hud.h
@@ -11,6 +11,6 @@ void drawHUD(PixelGrid *canvas, HUD *hud) {
     for (u32 i = 0; i < hud->line_count; i++, line++) {
         drawText(canvas, Color(line->title_color), line->title.char_ptr, x, y);
         drawText(canvas, Color(line->value_color), line->value.string.char_ptr, x + line->title.length * FONT_WIDTH, y);
-        y += hud->line_height * FONT_HEIGHT;
+        y += (i32)(hud->line_height * (f32)FONT_HEIGHT);
     }
 }
